refactor: extract AppState+PermissionRecovery (4 delegators) (#601 slice K)
Slice K of #601. Creates new AppState+PermissionRecovery.swift with 4
permissionRecoveryController delegators — 1 method, 2 computed properties,
and their supporting private helper.

Members:
- refreshPermissionRecoveryState() — method (single-expression: presenter.present(controller.refreshRecoveryState(...)))
- microphoneRecoveryGuidance — computed property (accesses private helper)
- microphoneRecoveryLocalTestingNote — computed property (accesses private helper)
- microphoneRecoveryGuidanceDetails — private helper (delegates to controller.microphoneRecoveryGuidance)

The private helper stays private within the new extension file (Swift's
file-scope private) — accessible only to the 2 public delegators that use it.
No visibility widen.

SHAs verified against origin/main:
- microphoneRecovery* properties (580-586): 49c6602cdc831e0ee978a801987e01bc39fb75ae1070455c6087258a4049a647
- refreshPermissionRecoveryState (599-606): e073c11a9e8ae577a7d273af6dafd2f6072bf2f0b1e875d6990b85e72a2bc48a
- private helper (1107-1109): 720bab1427d8f072b1cb66ac87550669481b996dac35c8f46ba9f737e696df24

Non-adjacent 3-block extraction: 580-586, 599-606, 1107-1109. Reverse-order sed per #556.

Safety checklist per #601:
- Delegator test: all pass (single-expression bodies, args observed from self or literal)
- $binding sanity check: rg '\$appState\.(microphoneRecoveryGuidance|microphoneRecoveryLocalTestingNote)' returned zero
- Observation forwarding preserved
- Green tests: 667
